chore(Scene): initial work for LibScene separation from engine library
