+++ dependecies.lua
@@ -15,13 +15,15 @@ Dependencies = {
         include = "%{wks.location}/vendor/GLFW/include",
         link = "glfw",
     },
+    entt = {
+        include = "%{wks.location}/vendor/entt/include",
+    },
 }
 IncludeDirectories = {}
 
 IncludeDirectories["imgui"] = "%{wks.location}/TerranEngine/vendor/ImGui/"
 IncludeDirectories["glad"] = "%{wks.location}/TerranEngine/vendor/GLAD/include/"
 IncludeDirectories["stb"] = "%{wks.location}/TerranEngine/vendor/stb/"
-IncludeDirectories["entt"] = "%{wks.location}/TerranEngine/vendor/entt/include/"
 IncludeDirectories["msdfgen"] = "%{wks.location}/TerranEngine/vendor/msdf-atlas-gen/msdfgen/"
 IncludeDirectories["msdf_atlas_gen"] = "%{wks.location}/TerranEngine/vendor/msdf-atlas-gen/"
 IncludeDirectories["box2d"] = "%{wks.location}/TerranEngine/vendor/Box2D/include/"
